Added some functions to OrderList and tested them
see the new directory created for the outputs testing the orderlist called OrderListTest.  Next step is adding all the functionality of the original bkoda and also figuring out how I want to organize the data for the cancelFail list and the first 300 etc.  It would be cool if the number of coolers wasn't hard coded etc.